requirements.txt
# Erguvan AI Content Generator - Production Dependencies

# Core AI/LLM libraries
openai==1.55.0
anthropic==0.40.0
langchain==0.3.8
langchain-openai==0.2.8
langchain-anthropic==0.2.4
langchain-community==0.3.8

# Vector storage and embeddings
faiss-cpu==1.9.0
sentence-transformers==3.3.1
numpy==1.26.4
scipy==1.13.1

# Document processing
python-docx==1.1.2
pdfplumber==0.11.4
python-pptx==1.0.2
markdown==3.7
beautifulsoup4==4.12.3
lxml==5.3.0
markitdown==0.0.1a1

# Web framework and API
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.0

# Database and persistence
sqlalchemy>=1.4,<2.0.36
alembic==1.14.0
# sqlite3 - Built-in with Python

# Template engine
jinja2==3.1.4

# Security and sanitization
bleach==6.2.0
python-magic==0.4.27

# Text processing and analysis
nltk==3.9.1
textstat==0.7.4
simhash==2.1.2
spacy==3.8.2

# Environment and configuration
python-dotenv==1.0.1
pyyaml==6.0.2
click==8.1.7

# Monitoring and logging
prometheus-client==0.21.1
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
httpx==0.28.1  # For testing FastAPI

# Code quality
black==24.10.0
ruff==0.8.4
mypy==1.13.0
pre-commit==4.0.1

# UI (optional)
streamlit==1.40.2
plotly==5.24.1

# Development utilities
python-dateutil==2.9.0.post0
tqdm==4.67.1
rich==13.9.4